dashboard_backend/pyproject.toml
[project]
name = "dashboard-backend"
version = "0.1.0"
description = "A minimal dashboard backend service"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi==0.115.6",
    "alembic==1.14.0",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "pyserial==3.5",
    "requests==2.32.3",
    "pymodbus==3.8.0",
    "socketio-client==0.7.2",
    "apscheduler==3.11.0",
    "python-socketio==5.11.4",
    "passlib==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "pydantic-settings==2.7.0",
    "bcrypt==4.2.1",
    "uvicorn",
    "httpx>=0.28.1",
]
requires-python = ">=3.11"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "ipython>=8.31.0",
    "nest-asyncio>=1.6.0",
    "pdbpp",
]
